+++ sdk/src/driftClient.ts
@@ -609,7 +609,8 @@ export class DriftClient {
 					(await this.addUser(account.subAccountId, account.authority));
 			}
 
-			if (this.activeSubAccountId == undefined) {
+			// sometimes delegate accounts load first, switch to first useraccount if it happens
+			if (this.activeSubAccountId == undefined || !this.activeAuthority.equals(this.authority)) {
 				this.switchActiveUser(
 					userAccounts.concat(delegatedAccounts)[0]?.subAccountId ?? 0,
 					userAccounts.concat(delegatedAccounts)[0]?.authority ?? this.authority