+++ sdk/src/index.ts
@@ -70,6 +70,7 @@ export * from './dlob/DLOBOrders';
 export * from './dlob/NodeList';
 export * from './dlob/DLOBSubscriber';
 export * from './dlob/types';
+export * from './dlob/orderBookLevels';
 export * from './userMap/userMap';
 export * from './userMap/userStatsMap';
 export * from './math/bankruptcy';